+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.1
 
 - Implement FluentLogger\#post
+- Implement retry connection when connection failed